fix: fix some unit tests not passing
Some unit tests were failing with the following message:
“A Timer is still pending even after the widget tree was disposed.”

The cause of that was that `VisualEffectSubviewContainer`, or more
precisely `VisualEffectSubviewContainerWithGlobalKey` is using a timer
to update its visual effect subview outside of the widget's `build`
method.

The issue is fixed by either disabling wallpaper tinting in the window
(and therefore eliminating the use of `VisualEffectSubviewContainer`)
or, in cases where that was impossible (such as the sidebar) running
`await tester.pump(Duration.zero);` to allow the timer to complete.